+++ CHANGELOG-V3.md
@@ -1,0 +1,2 @@
+### Removed
+- `utils.filter_params`
