+++ discord/iterators.py
@@ -41,7 +41,7 @@ from typing import (
 from .audit_logs import AuditLogEntry
 from .errors import NoMoreItems
 from .object import Object
-from .utils import maybe_coroutine, snowflake_time, time_snowflake
+from .utils import maybe_coroutine, snowflake_time, generate_snowflake
 
 __all__ = (
     "ReactionIterator",
@@ -341,11 +341,11 @@ class HistoryIterator(_AsyncIterator["Message"]):
         oldest_first=None,
     ):
         if isinstance(before, datetime.datetime):
-            before = Object(id=time_snowflake(before, high=False))
+            before = Object(id=generate_snowflake(before, high=False))
         if isinstance(after, datetime.datetime):
-            after = Object(id=time_snowflake(after, high=True))
+            after = Object(id=generate_snowflake(after, high=True))
         if isinstance(around, datetime.datetime):
-            around = Object(id=time_snowflake(around))
+            around = Object(id=generate_snowflake(around))
 
         self.reverse = after is not None if oldest_first is None else oldest_first
         self.messageable = messageable
@@ -467,9 +467,9 @@ class AuditLogIterator(_AsyncIterator["AuditLogEntry"]):
         action_type=None,
     ):
         if isinstance(before, datetime.datetime):
-            before = Object(id=time_snowflake(before, high=False))
+            before = Object(id=generate_snowflake(before, high=False))
         if isinstance(after, datetime.datetime):
-            after = Object(id=time_snowflake(after, high=True))
+            after = Object(id=generate_snowflake(after, high=True))
 
         self.guild = guild
         self.loop = guild._state.loop
@@ -578,9 +578,9 @@ class GuildIterator(_AsyncIterator["Guild"]):
 
     def __init__(self, bot, limit, before=None, after=None, with_counts=True):
         if isinstance(before, datetime.datetime):
-            before = Object(id=time_snowflake(before, high=False))
+            before = Object(id=generate_snowflake(before, high=False))
         if isinstance(after, datetime.datetime):
-            after = Object(id=time_snowflake(after, high=True))
+            after = Object(id=generate_snowflake(after, high=True))
 
         self.bot = bot
         self.limit = limit
@@ -665,7 +665,7 @@ class GuildIterator(_AsyncIterator["Guild"]):
 class MemberIterator(_AsyncIterator["Member"]):
     def __init__(self, guild, limit=1000, after=None):
         if isinstance(after, datetime.datetime):
-            after = Object(id=time_snowflake(after, high=True))
+            after = Object(id=generate_snowflake(after, high=True))
 
         self.guild = guild
         self.limit = limit
@@ -797,7 +797,7 @@ class ArchivedThreadIterator(_AsyncIterator["Thread"]):
             self.before = None
         elif isinstance(before, datetime.datetime):
             if joined:
-                self.before = str(time_snowflake(before, high=False))
+                self.before = str(generate_snowflake(before, high=False))
             else:
                 self.before = before.isoformat()
         else:
@@ -873,9 +873,9 @@ class ScheduledEventSubscribersIterator(_AsyncIterator[Union["User", "Member"]])
         after: datetime.datetime | int | None = None,
     ):
         if isinstance(before, datetime.datetime):
-            before = Object(id=time_snowflake(before, high=False))
+            before = Object(id=generate_snowflake(before, high=False))
         if isinstance(after, datetime.datetime):
-            after = Object(id=time_snowflake(after, high=True))
+            after = Object(id=generate_snowflake(after, high=True))
 
         self.event = event
         self.limit = limit
@@ -967,9 +967,9 @@ class EntitlementIterator(_AsyncIterator["Entitlement"]):
         self.sku_ids = sku_ids
 
         if isinstance(before, datetime.datetime):
-            before = Object(id=time_snowflake(before, high=False))
+            before = Object(id=generate_snowflake(before, high=False))
         if isinstance(after, datetime.datetime):
-            after = Object(id=time_snowflake(after, high=True))
+            after = Object(id=generate_snowflake(after, high=True))
 
         self.before = before
         self.after = after
@@ -1081,9 +1081,9 @@ class SubscriptionIterator(_AsyncIterator["Subscription"]):
         user_id: int | None = None,
     ):
         if isinstance(before, datetime.datetime):
-            before = Object(id=time_snowflake(before, high=False))
+            before = Object(id=generate_snowflake(before, high=False))
         if isinstance(after, datetime.datetime):
-            after = Object(id=time_snowflake(after, high=True))
+            after = Object(id=generate_snowflake(after, high=True))
 
         self.state = state
         self.sku_id = sku_id
