+++ discord/onboarding.py
@@ -81,7 +81,7 @@ class PromptOption:
         id: int | None = None,
     ):
         # ID is required when making edits, but it can be any snowflake that isn't already used by another prompt during edits
-        self.id: int = int(id) if id else generate_snowflake()
+        self.id: int = int(id) if id else generate_snowflake(mode="realistic")
         self.title: str = title
         self.channels: list[Snowflake] = channels or []
         self.roles: list[Snowflake] = roles or []
@@ -169,7 +169,7 @@ class OnboardingPrompt:
         id: int | None = None,  # Currently optional as users can manually create these
     ):
         # ID is required when making edits, but it can be any snowflake that isn't already used by another prompt during edits
-        self.id: int = int(id) if id else generate_snowflake()
+        self.id: int = int(id) if id else generate_snowflake(mode="realistic")
 
         self.type: PromptType = type
         if isinstance(self.type, int):
