+++ discord/utils/__init__.py
@@ -64,7 +64,9 @@ from typing import (
     overload,
 )
 
-from .errors import HTTPException, InvalidArgument
+from ..errors import HTTPException, InvalidArgument
+
+from .public import basic_autocomplete, generate_snowflake, utcnow
 
 try:
     import msgspec
@@ -80,7 +82,6 @@ __all__ = (
     "deprecated",
     "oauth_url",
     "snowflake_time",
-    "time_snowflake",
     "find",
     "get",
     "get_or_fetch",
@@ -98,10 +99,7 @@ __all__ = (
     "format_dt",
     "generate_snowflake",
     "basic_autocomplete",
-    "filter_params",
 )
-
-DISCORD_EPOCH = 1420070400000
 
 
 class Undefined(Enum):
@@ -444,32 +442,6 @@ def snowflake_time(id: int) -> datetime.datetime:
     return datetime.datetime.fromtimestamp(timestamp, tz=datetime.timezone.utc)
 
 
-def time_snowflake(dt: datetime.datetime, high: bool = False) -> int:
-    """Returns a numeric snowflake pretending to be created at the given date.
-
-    When using as the lower end of a range, use ``time_snowflake(high=False) - 1``
-    to be inclusive, ``high=True`` to be exclusive.
-
-    When using as the higher end of a range, use ``time_snowflake(high=True) + 1``
-    to be inclusive, ``high=False`` to be exclusive
-
-    Parameters
-    ----------
-    dt: :class:`datetime.datetime`
-        A datetime object to convert to a snowflake.
-        If naive, the timezone is assumed to be local time.
-    high: :class:`bool`
-        Whether to set the lower 22 bit to high or low.
-
-    Returns
-    -------
-    :class:`int`
-        The snowflake representing the time given.
-    """
-    discord_millis = int(dt.timestamp() * 1000 - DISCORD_EPOCH)
-    return (discord_millis << 22) + (2**22 - 1 if high else 0)
-
-
 def find(predicate: Callable[[T], Any], seq: Iterable[T]) -> T | None:
     """A helper to return the first element found in the sequence
     that meets the predicate. For example: ::
@@ -740,22 +712,6 @@ async def sleep_until(when: datetime.datetime, result: T | None = None) -> T | N
     """
     delta = compute_timedelta(when)
     return await asyncio.sleep(delta, result)
-
-
-def utcnow() -> datetime.datetime:
-    """A helper function to return an aware UTC datetime representing the current time.
-
-    This should be preferred to :meth:`datetime.datetime.utcnow` since it is an aware
-    datetime, compared to the naive datetime in the standard library.
-
-    .. versionadded:: 2.0
-
-    Returns
-    -------
-    :class:`datetime.datetime`
-        The current aware datetime in UTC.
-    """
-    return datetime.datetime.now(datetime.timezone.utc)
 
 
 def valid_icon_size(size: int) -> bool:
@@ -1243,7 +1199,7 @@ def format_dt(dt: datetime.datetime | datetime.time, /, style: TimestampStyle | 
     ----------
     dt: Union[:class:`datetime.datetime`, :class:`datetime.time`]
         The datetime to format.
-    style: :class:`str`
+    style: :class:`str`R
         The style to format the datetime with.
 
     Returns
@@ -1256,118 +1212,3 @@ def format_dt(dt: datetime.datetime | datetime.time, /, style: TimestampStyle | 
     if style is None:
         return f"<t:{int(dt.timestamp())}>"
     return f"<t:{int(dt.timestamp())}:{style}>"
-
-
-def generate_snowflake(dt: datetime.datetime | None = None) -> int:
-    """Returns a numeric snowflake pretending to be created at the given date but more accurate and random
-    than :func:`time_snowflake`. If dt is not passed, it makes one from the current time using utcnow.
-
-    Parameters
-    ----------
-    dt: :class:`datetime.datetime`
-        A datetime object to convert to a snowflake.
-        If naive, the timezone is assumed to be local time.
-
-    Returns
-    -------
-    :class:`int`
-        The snowflake representing the time given.
-    """
-
-    dt = dt or utcnow()
-    return int(dt.timestamp() * 1000 - DISCORD_EPOCH) << 22 | 0x3FFFFF
-
-
-V = Union[Iterable[OptionChoice], Iterable[str], Iterable[int], Iterable[float]]
-AV = Awaitable[V]
-Values = Union[V, Callable[[AutocompleteContext], Union[V, AV]], AV]
-AutocompleteFunc = Callable[[AutocompleteContext], AV]
-FilterFunc = Callable[[AutocompleteContext, Any], Union[bool, Awaitable[bool]]]
-
-
-def basic_autocomplete(values: Values, *, filter: FilterFunc | None = None) -> AutocompleteFunc:
-    """A helper function to make a basic autocomplete for slash commands. This is a pretty standard autocomplete and
-    will return any options that start with the value from the user, case-insensitive. If the ``values`` parameter is
-    callable, it will be called with the AutocompleteContext.
-
-    This is meant to be passed into the :attr:`discord.Option.autocomplete` attribute.
-
-    Parameters
-    ----------
-    values: Union[Union[Iterable[:class:`.OptionChoice`], Iterable[:class:`str`], Iterable[:class:`int`], Iterable[:class:`float`]], Callable[[:class:`.AutocompleteContext`], Union[Union[Iterable[:class:`str`], Iterable[:class:`int`], Iterable[:class:`float`]], Awaitable[Union[Iterable[:class:`str`], Iterable[:class:`int`], Iterable[:class:`float`]]]]], Awaitable[Union[Iterable[:class:`str`], Iterable[:class:`int`], Iterable[:class:`float`]]]]
-        Possible values for the option. Accepts an iterable of :class:`str`, a callable (sync or async) that takes a
-        single argument of :class:`.AutocompleteContext`, or a coroutine. Must resolve to an iterable of :class:`str`.
-    filter: Optional[Callable[[:class:`.AutocompleteContext`, Any], Union[:class:`bool`, Awaitable[:class:`bool`]]]]
-        An optional callable (sync or async) used to filter the autocomplete options. It accepts two arguments:
-        the :class:`.AutocompleteContext` and an item from ``values`` iteration treated as callback parameters. If ``None`` is provided, a default filter is used that includes items whose string representation starts with the user's input value, case-insensitive.
-
-        .. versionadded:: 2.7
-
-    Returns
-    -------
-    Callable[[:class:`.AutocompleteContext`], Awaitable[Union[Iterable[:class:`.OptionChoice`], Iterable[:class:`str`], Iterable[:class:`int`], Iterable[:class:`float`]]]]
-        A wrapped callback for the autocomplete.
-
-    Examples
-    --------
-
-    Basic usage:
-
-    .. code-block:: python3
-
-        Option(str, "color", autocomplete=basic_autocomplete(("red", "green", "blue")))
-
-        # or
-
-
-        async def autocomplete(ctx):
-            return "foo", "bar", "baz", ctx.interaction.user.name
-
-
-        Option(str, "name", autocomplete=basic_autocomplete(autocomplete))
-
-    With filter parameter:
-
-    .. code-block:: python3
-
-        Option(
-            str,
-            "color",
-            autocomplete=basic_autocomplete(("red", "green", "blue"), filter=lambda c, i: str(c.value or "") in i),
-        )
-
-    .. versionadded:: 2.0
-
-    Note
-    ----
-    Autocomplete cannot be used for options that have specified choices.
-    """
-
-    async def autocomplete_callback(ctx: AutocompleteContext) -> V:
-        _values = values  # since we reassign later, python considers it local if we don't do this
-
-        if callable(_values):
-            _values = _values(ctx)
-        if asyncio.iscoroutine(_values):
-            _values = await _values
-
-        if filter is None:
-
-            def _filter(ctx: AutocompleteContext, item: Any) -> bool:
-                item = getattr(item, "name", item)
-                return str(item).lower().startswith(str(ctx.value or "").lower())
-
-            gen = (val for val in _values if _filter(ctx, val))
-
-        elif asyncio.iscoroutinefunction(filter):
-            gen = (val for val in _values if await filter(ctx, val))
-
-        elif callable(filter):
-            gen = (val for val in _values if filter(ctx, val))
-
-        else:
-            raise TypeError("``filter`` must be callable.")
-
-        return iter(itertools.islice(gen, 25))
-
-    return autocomplete_callback
