+++ pyproject.toml
@@ -75,6 +75,7 @@ dev = [
     "pylint~=3.3.6",
     "pytest~=8.3.5",
     "pytest-asyncio~=0.24.0",
+    "python-dotenv>=1.1.0",
     "ruff>=0.11.9",
 ]
 
@@ -100,7 +101,7 @@ artifacts = ["discord/bin/*.dll"]
 target-version = ['py310', 'py311', 'py312', 'py313']
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py310"
 line-length = 120
 indent-width = 4
